pyproject.toml
[tool.poetry]
name = "Hello"
version = "1.1"
description = ""
authors = ["bbalandin <balandin_bogdan@list.ru>"]

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.19.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
